requirements.txt
# requirements.txt

fastapi
uvicorn[standard]
pydantic
requests
beautifulsoup4
google-genai
python-dotenv